Allow easier overriding of LDAPBackend.authenticate() method
Use case:

After a successful LDAP authenticate, a user may still not be ready to save to
the Django database without additional information. My project has a workflow
for such an event. I must serialize the available LDAP attributes to a storage
before redirecting to this workflow. The current implementation of
LDAPBackend.authenticate() makes it difficult to access the LDAP attributes
without either rewriting the same logic or rebuilding the _LDAPUser object. I'd
like to avoid both of these.