Merged in jdufresne/django-auth-ldap (pull request #71)
Use environment markers to describe requirements

The environment markers allow specifying dependencies for different
environments. For example, different dependencies for Python2 or Python3. Using
environment markers has better support among tools than calculating
dependencies during setup.py execution. For full details on environment
markers, see PEP 508.

https://www.python.org/dev/peps/pep-0508/

This allows for a stable and static set of dependencies. For example, currently
when requesting the PyPI information with the URL:

https://pypi.python.org/pypi/django-auth-ldap/json

Only python-ldap is listed as a requirement. After this change, all
dependencies will be listed with their respective markers.

As the list of dependencies is now static and stable, can enable universal
wheel support. This will allow the package to be distributed as a single wheel
instead of two. For more information on wheels, see:

https://pythonwheels.com/
https://www.python.org/dev/peps/pep-0427/