Restore use of isort for sorting Python imports
When black was introduced in commit
5c02dc2b5e272efd7619dcbcfe335a71c58d5261, isort was removed. It was
removed under the mistaken impression that black would sort imports
making isort redundant. However, that is not the case. Black has a
policy of only applying format changes and never changing the Python
AST. Reordering imports is altering the AST as it effects execution
order. Restore isort to once again apply a consistent, predictable
Python import order. The isort configuration was adopted from the black
README.md.